Moving things around to fix some circular references that are directly within our current package.
