+++ sdk/eventhub/event-hubs/src/eventHubConsumerClientModels.ts
@@ -1,4 +1,4 @@
-import { CloseReason } from "./eventProcessor";
+import { CloseReason } from "./models.public";
 import { ReceivedEventData } from "./eventData";
 import { LastEnqueuedEventProperties } from "./eventHubReceiver";
 import { EventPosition } from "./eventPosition";
@@ -90,7 +90,7 @@ export type ProcessErrorHandler = (
 ) => Promise<void>;
 
 /**
- * Signature of the user provided function invoked by `EventHubConsumerClient` just before starting to receive 
+ * Signature of the user provided function invoked by `EventHubConsumerClient` just before starting to receive
  * events from a partition.
  */
 export type ProcessInitializeHandler = (context: PartitionContext) => Promise<void>;
@@ -110,7 +110,7 @@ export interface SubscriptionEventHandlers {
   /**
    * The function invoked by `EventHubConsumerClient` when a set of events is received. The
    * `PartitionContext` passed to this function can be used to determine which partition is being read from.
-   * 
+   *
    * The `updateCheckpoint()` method on the context can be used to update checkpoints in the `CheckpointStore`
    * (if one was provided to the client). Use this in frequent intervals to mark events that have been processed
    * so that the client can restart from such checkpoints in the event of a restart or error recovery.
@@ -119,19 +119,19 @@ export interface SubscriptionEventHandlers {
   /**
    * The function invoked by `EventHubConsumerClient` for errors that occur when receiving events
    * or when executing any of the user provided functions passed to the `subscribe()` method.
-   * 
+   *
    * The `PartitionContext` passed to this function will indicate the partition that was being processed
    * when the error was thrown. In cases where an error is thrown outside of processing events from a
    * partition(e.g. failure to perform load balancing), the `partitionId` on the context will be an empty string.
-   * 
+   *
    * After the client completes executing this function, the `partitionClose` function is invoked.
    */
   processError: ProcessErrorHandler;
   /**
    * The function invoked by `EventHubConsumerClient` each time the subscription is about to begin
-   * reading from a partition. The `PartitionContext` passed to this function can be used to determine 
+   * reading from a partition. The `PartitionContext` passed to this function can be used to determine
    * which partition is about to be read from.
-   * 
+   *
    * The client will start receiving events for the partition only after completing the execution of
    * this function (if provided). Therefore, use this function to carry out any setup work including
    * async tasks.
@@ -139,13 +139,13 @@ export interface SubscriptionEventHandlers {
   processInitialize?: ProcessInitializeHandler;
   /**
    * The function invoked by `EventHubConsumerClient` each time the subscription stops reading events from
-   * a partition. The information on this partition will be available on the `PartitionContext` passed to the 
+   * a partition. The information on this partition will be available on the `PartitionContext` passed to the
    * function `processClose`.
-   * 
+   *
    * If the `CloseReason` passed to this function is `OwnershipLost`, then another subscription has taken over
    * reading from the same partition using the same consumer group. This is expected if you have multiple
    * instances of your application running and have passed the `CheckpointStore` to the client to load balance.
-   * 
+   *
    * If the `CloseReason` is `Shutdown`, this indicates that either `subscription.close()` was called, or an
    * error occured. Unless the subscription was explicitly closed via `subscription.close()`, the subscription
    * will attempt to resume reading events from the last checkpoint for the partition.
