+++ sdk/eventhub/event-hubs/src/eventHubSender.ts
@@ -25,11 +25,10 @@ import {
 import { EventData, toAmqpMessage } from "./eventData";
 import { ConnectionContext } from "./connectionContext";
 import { LinkEntity } from "./linkEntity";
-import {
-  SendOptions,
-  EventHubProducerOptions,
-  getRetryAttemptTimeoutInMs
-} from "./impl/eventHubClient";
+import { EventHubProducerOptions } from "./impl/models.private";
+import { SendOptions } from "./models.public";
+
+import { getRetryAttemptTimeoutInMs } from "./util/retries";
 import { AbortSignalLike, AbortError } from "@azure/abort-controller";
 import { EventDataBatch, isEventDataBatch } from "./eventDataBatch";
