+++ sdk/eventhub/event-hubs/src/index.ts
@@ -15,7 +15,7 @@ export {
   GetPartitionPropertiesOptions,
   GetEventHubPropertiesOptions,
   WebSocketOptions
-} from "./impl/eventHubClient";
+} from "./models.public";
 export { EventHubConsumerClient } from "./eventHubConsumerClient";
 export { EventHubProducerClient } from "./eventHubProducerClient";
 export {
@@ -32,6 +32,7 @@ export { EventPosition, latestEventPosition, earliestEventPosition } from "./eve
 export { PartitionProperties, EventHubProperties } from "./managementClient";
 export { EventDataBatch, TryAddOptions } from "./eventDataBatch";
 export { Checkpoint } from "./partitionProcessor";
-export { CloseReason, CheckpointStore, PartitionOwnership } from "./eventProcessor";
+export { CheckpointStore, PartitionOwnership } from "./eventProcessor";
+export { CloseReason } from "./models.public";
 export { MessagingError, RetryOptions, TokenCredential } from "@azure/core-amqp";
 export { logger } from "./log";
