+++ sdk/eventhub/event-hubs/src/managementClient.ts
@@ -25,7 +25,7 @@ import {
 import { ConnectionContext } from "./connectionContext";
 import { LinkEntity } from "./linkEntity";
 import { logger, logErrorStackTrace } from "./log";
-import { getRetryAttemptTimeoutInMs } from "./impl/eventHubClient";
+import { getRetryAttemptTimeoutInMs } from "./util/retries";
 import { AbortSignalLike, AbortError } from "@azure/abort-controller";
 /**
  * Describes the runtime information of an Event Hub.
