+++ sdk/eventhub/event-hubs/src/partitionProcessor.ts
@@ -1,4 +1,5 @@
-import { CloseReason, CheckpointStore } from "./eventProcessor";
+import { CheckpointStore } from "./eventProcessor";
+import { CloseReason } from "./models.public";
 import { ReceivedEventData } from "./eventData";
 import { LastEnqueuedEventProperties } from "./eventHubReceiver";
 import {
