+++ sdk/eventhub/event-hubs/src/pumpManager.ts
@@ -3,7 +3,8 @@
 
 import { EventHubClient } from "./impl/eventHubClient";
 import { EventPosition } from "./eventPosition";
-import { FullEventProcessorOptions, CloseReason } from "./eventProcessor";
+import { CommonEventProcessorOptions } from "./impl/models.private";
+import { CloseReason } from "./models.public";
 import { PartitionProcessor } from "./partitionProcessor";
 import { PartitionPump } from "./partitionPump";
 import { logger, logErrorStackTrace } from "./log";
@@ -55,7 +56,7 @@ export interface PumpManager {
  */
 export class PumpManagerImpl implements PumpManager {
   private readonly _eventProcessorName: string;
-  private readonly _options: FullEventProcessorOptions;
+  private readonly _options: CommonEventProcessorOptions;
   private _partitionIdToPumps: {
     [partitionId: string]: PartitionPump | undefined;
   } = {};
@@ -63,7 +64,7 @@ export class PumpManagerImpl implements PumpManager {
   /**
    * @ignore
    */
-  constructor(eventProcessorName: string, eventProcessorOptions: FullEventProcessorOptions) {
+  constructor(eventProcessorName: string, eventProcessorOptions: CommonEventProcessorOptions) {
     this._eventProcessorName = eventProcessorName;
     this._options = eventProcessorOptions;
   }
