+++ sdk/eventhub/event-hubs/test/sender.spec.ts
@@ -8,7 +8,8 @@ chai.use(chaiAsPromised);
 import debugModule from "debug";
 const debug = debugModule("azure:event-hubs:sender-spec");
 import { EventData, EventHubProducerClient, EventHubConsumerClient } from "../src";
-import { SendOptions, EventHubClient } from "../src/impl/eventHubClient";
+import { EventHubClient } from "../src/impl/eventHubClient";
+import { SendOptions } from "../src/models.public";
 import { EnvVarKeys, getEnvVars } from "./utils/testUtils";
 import { AbortController } from "@azure/abort-controller";
 import { TestTracer, setTracer, SpanGraph } from "@azure/core-tracing";
